% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{Simple Tree Data 1}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{from}{A node label and started point of directed edge (parent node)}
 \item{to}{A node label and endpoint of directed edge (child node)}
 \item{Estimate}{A numerical value assumed to be survey count belonging to 'to' node (integer)}
 \item{Total}{A numerical value assumed to be survey sample size (integer)}
 \item{Count}{A numerical value for marginal count if leaf node (integer)}
 \item{Population}{A boolean value for if survey size is entire population (logical)}
 \item{Description}{A string describing 'to' node (string)}
}
}
\usage{
data(data1)
}
\description{
Small, artificially generated toy data set to demonstrate package functionality
}
\examples{
data(data1)
head(data1)

}
\references{
This data set was artificially created for the JAGStree package.
}
\keyword{datasets}
