% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{Simple Tree Data 2}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{from}{A node label and started point of directed edge (parent node)}
 \item{to}{A node label and endpoint of directed edge (child node)}
 }
}
\usage{
data(data2)
}
\description{
Small, artificially generated toy data set to demonstrate package functionality
with a simpler data frame
}
\examples{
data(data2)
head(data2)

}
\references{
This data set was artificially created for the JAGStree package.
}
\keyword{datasets}
